+++ log-collector/src/test/java/core/log/web/EventControllerTest.java
@@ -1,16 +1,23 @@
 package core.log.web;
 
+import core.framework.inject.Inject;
+import core.framework.json.Bean;
 import core.framework.log.message.EventMessage;
+import core.framework.util.Strings;
 import core.framework.web.Request;
 import core.framework.web.exception.ForbiddenException;
+import core.log.IntegrationTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.time.Instant;
+import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Optional;
 import java.util.Set;
 
+import static core.log.web.SendEventRequest.Event;
+import static core.log.web.SendEventRequest.Result;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.mock;
@@ -19,7 +26,9 @@ import static org.mockito.Mockito.when;
 /**
  * @author neo
  */
-class EventControllerTest {
+class EventControllerTest extends IntegrationTest {
+    @Inject
+    SendEventRequestValidator validator;
     private Request request;
 
     @BeforeEach
@@ -61,9 +70,9 @@ class EventControllerTest {
 
     @Test
     void message() {
-        var event = new SendEventRequest.Event();
+        var event = new Event();
         event.date = ZonedDateTime.now().minusHours(1);
-        event.result = SendEventRequest.Result.WARN;
+        event.result = Result.WARN;
         event.errorCode = "NOT_FOUND";
         event.errorMessage = "not found";
         event.context.put("path", "/path");
@@ -85,5 +94,24 @@ class EventControllerTest {
         assertThat(message.info).isEqualTo(event.info);
         assertThat(message.elapsed).isEqualTo(event.elapsedTime);
         assertThat(message.app).isEqualTo("test");
+    }
+
+    @Test
+    void sendEventRequest() {
+        var event = new Event();
+        event.date = ZonedDateTime.now().withZoneSameInstant(ZoneId.of("UTC"));
+        event.result = Result.OK;
+        event.action = "action";
+        event.elapsedTime = 10L;
+        var sendEventRequest = new SendEventRequest();
+        sendEventRequest.events.add(event);
+
+        when(request.body()).thenReturn(Optional.of(Strings.bytes(Bean.toJSON(sendEventRequest))));
+
+        var controller = new EventController(Set.of());
+        controller.validator = validator;
+        SendEventRequest parsedSendEventRequest = controller.sendEventRequest(request);
+
+        assertThat(parsedSendEventRequest).usingRecursiveComparison().isEqualTo(sendEventRequest);
     }
 }